All functional

+++ app/src/main/java/com/netjob/habittracker/data/HabitDataContract.java
@@ -2,7 +2,6 @@ package com.netjob.habittracker.data;
 
 import android.provider.BaseColumns;
 
-import com.netjob.habittracker.R;
 
 /**
  * Created by root on 1/26/17.
@@ -16,14 +15,13 @@ public class HabitDataContract {
 
         private HabitEntry() {}
 
-
         public static final String TABLE_NAME = "habits";
 
         public static final int RAN_EQUALS_YES = 1;
         public static final int RAN_EQUALS_NO = 0;
 
         public static final int RAN_DEFAULT = RAN_EQUALS_NO;
-        public static final int MILES_DEFAULT = 0;
+        public static final double MILES_DEFAULT = 0.0;
         public static final int PUSHUPS_DEFAULT = 0;
         public static final int WATERCUPS_DEFAULT = 0;
         public static final String SYMPTOMS_DEFAULT = "none";
@@ -34,9 +32,6 @@ public class HabitDataContract {
         public static final String COLUMN_NAME_SYMPTOMS = "symptoms";
         public static final String COLUMN_NAME_RAN = "ran";
         public static final String COLUMN_NAME_MILES = "miles";
-
-
-
 
     }
